Standardize spec python-oslo-i18n
Spec for oslo-i18n differs from the rest of oslo packages.
This makes difficult to apply some changes we are doing in
other packages as i18n cleanup.

This review applies a quite standard spec for i18n. Note it
still has different -doc packages for py3 and py2 but i think
we can change that in a later review if required.

Change-Id: I91f3888ee962b5c24d1219523a70e6cc8fcf4664